glsl template: fix iFrame datatype

+++ shady-app/src/template.glsl
@@ -11,7 +11,7 @@ layout(binding = 1) uniform vec2 iResolution;
 // z: treble
 layout(binding = 2) uniform vec3 iAudio;
 
-layout(binding = 4) uniform double iFrame;
+layout(binding = 4) uniform uint iFrame;
 
 layout(location = 0) out vec4 fragColor;
